refactor: move test code to shared
